Move DefaultValuesFromSchema to pkgUtils
