+++ cmd/kubeapps-apis/plugins/pkg/pkgutils/pkgutils.go
@@ -7,7 +7,10 @@
 package pkgutils
 
 import (
+	"bufio"
+	"fmt"
 	"net/url"
+	"reflect"
 	"strings"
 
 	"github.com/Masterminds/semver/v3"
@@ -16,6 +19,10 @@ import (
 	"github.com/kubeapps/kubeapps/pkg/chart/models"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
+	"gopkg.in/yaml.v3" // The usual "sigs.k8s.io/yaml" doesn't work: https://github.com/kubeapps/kubeapps/pull/4050
+	"k8s.io/apiextensions-apiserver/pkg/apis/apiextensions"
+	structuralschema "k8s.io/apiextensions-apiserver/pkg/apiserver/schema"
+	"k8s.io/apimachinery/pkg/runtime"
 )
 
 // Contains miscellaneous package-utilities used by multiple plug-ins
@@ -198,4 +205,174 @@ func SplitChartIdentifier(chartID string) (repoName, chartName string, err error
 	}
 	chartIDParts := strings.Split(unescapedChartID, "/")
 	return chartIDParts[0], chartIDParts[1], nil
+}
+
+// DefaultValuesFromSchema returns a yaml string with default values generated from an OpenAPI v3 Schema
+func DefaultValuesFromSchema(schema []byte, isCommentedOut bool) (string, error) {
+	if len(schema) == 0 {
+		return "", nil
+	}
+	// Deserialize the schema passed into the function
+	jsonSchemaProps := &apiextensions.JSONSchemaProps{}
+	if err := yaml.Unmarshal(schema, jsonSchemaProps); err != nil {
+		return "", err
+	}
+	structural, err := structuralschema.NewStructural(jsonSchemaProps)
+	if err != nil {
+		return "", err
+	}
+
+	// Generate the default values
+	unstructuredDefaultValues := make(map[string]interface{})
+	defaultValues(unstructuredDefaultValues, structural)
+	yamlDefaultValues, err := yaml.Marshal(unstructuredDefaultValues)
+	if err != nil {
+		return "", err
+	}
+	strYamlDefaultValues := string(yamlDefaultValues)
+
+	// If isCommentedOut, add a yaml comment character '#' to the beginning of each line
+	if isCommentedOut {
+		var sb strings.Builder
+		scanner := bufio.NewScanner(strings.NewReader(strYamlDefaultValues))
+		scanner.Split(bufio.ScanLines)
+		for scanner.Scan() {
+			sb.WriteString("# ")
+			sb.WriteString(fmt.Sprintln(scanner.Text()))
+		}
+		strYamlDefaultValues = sb.String()
+	}
+	return strYamlDefaultValues, nil
+}
+
+// defaultValuesFromSchema returns a yaml string with default values generated from an OpenAPI v3 Schema
+func defaultValuesFromSchema(schema []byte, isCommentedOut bool) (string, error) {
+	if len(schema) == 0 {
+		return "", nil
+	}
+	// Deserialize the schema passed into the function
+	jsonSchemaProps := &apiextensions.JSONSchemaProps{}
+	if err := yaml.Unmarshal(schema, jsonSchemaProps); err != nil {
+		return "", err
+	}
+	structural, err := structuralschema.NewStructural(jsonSchemaProps)
+	if err != nil {
+		return "", err
+	}
+
+	// Generate the default values
+	unstructuredDefaultValues := make(map[string]interface{})
+	defaultValues(unstructuredDefaultValues, structural)
+	yamlDefaultValues, err := yaml.Marshal(unstructuredDefaultValues)
+	if err != nil {
+		return "", err
+	}
+	strYamlDefaultValues := string(yamlDefaultValues)
+
+	// If isCommentedOut, add a yaml comment character '#' to the beginning of each line
+	if isCommentedOut {
+		var sb strings.Builder
+		scanner := bufio.NewScanner(strings.NewReader(strYamlDefaultValues))
+		scanner.Split(bufio.ScanLines)
+		for scanner.Scan() {
+			sb.WriteString("# ")
+			sb.WriteString(fmt.Sprintln(scanner.Text()))
+		}
+		strYamlDefaultValues = sb.String()
+	}
+	return strYamlDefaultValues, nil
+}
+
+// Default does defaulting of x depending on default values in s.
+// Based upon https://github.com/kubernetes/apiextensions-apiserver/blob/release-1.21/pkg/apiserver/schema/defaulting/algorithm.go
+// Plus modifications from https://github.com/vmware-tanzu/tanzu-framework/pull/1422
+// In short, it differs from upstream in that:
+// -- 1. Prevent deep copy of int as it panics
+// -- 2. For type object scan the first level properties for any defaults to create an empty map to populate
+// -- 3. If the property does not have a default, add one based on the type ("", false, etc)
+func defaultValues(x interface{}, s *structuralschema.Structural) {
+	if s == nil {
+		return
+	}
+
+	switch x := x.(type) {
+	case map[string]interface{}:
+		for k, prop := range s.Properties { //nolint
+			// if Default for object is nil, scan first level of properties for any defaults to create an empty default
+			if prop.Default.Object == nil {
+				createDefault := false
+				if prop.Properties != nil {
+					for _, v := range prop.Properties { //nolint
+						if v.Default.Object != nil {
+							createDefault = true
+							break
+						}
+					}
+				}
+				if createDefault {
+					prop.Default.Object = make(map[string]interface{})
+					// If not generating an empty object, fall back to the data type's defaults
+				} else {
+					switch prop.Type {
+					case "string":
+						prop.Default.Object = ""
+					case "number":
+						prop.Default.Object = 0
+					case "integer":
+						prop.Default.Object = 0
+					case "boolean":
+						prop.Default.Object = false
+					case "array":
+						prop.Default.Object = []interface{}{}
+					case "object":
+						prop.Default.Object = make(map[string]interface{})
+					}
+				}
+			}
+			if _, found := x[k]; !found || isNonNullableNull(x[k], &prop) {
+				if isKindInt(prop.Default.Object) {
+					x[k] = prop.Default.Object
+				} else {
+					x[k] = runtime.DeepCopyJSONValue(prop.Default.Object)
+				}
+			}
+		}
+		for k := range x {
+			if prop, found := s.Properties[k]; found {
+				defaultValues(x[k], &prop)
+			} else if s.AdditionalProperties != nil {
+				if isNonNullableNull(x[k], s.AdditionalProperties.Structural) {
+					if isKindInt(s.AdditionalProperties.Structural.Default.Object) {
+						x[k] = s.AdditionalProperties.Structural.Default.Object
+					} else {
+						x[k] = runtime.DeepCopyJSONValue(s.AdditionalProperties.Structural.Default.Object)
+					}
+				}
+				defaultValues(x[k], s.AdditionalProperties.Structural)
+			}
+		}
+	case []interface{}:
+		for i := range x {
+			if isNonNullableNull(x[i], s.Items) {
+				if isKindInt(s.Items.Default.Object) {
+					x[i] = s.Items.Default.Object
+				} else {
+					x[i] = runtime.DeepCopyJSONValue(s.Items.Default.Object)
+				}
+			}
+			defaultValues(x[i], s.Items)
+		}
+	default:
+		// scalars, do nothing
+	}
+}
+
+// isNonNullalbeNull returns true if the item is nil AND it's nullable
+func isNonNullableNull(x interface{}, s *structuralschema.Structural) bool {
+	return x == nil && s != nil && !s.Generic.Nullable
+}
+
+// isKindInt returns true if the item is an int
+func isKindInt(src interface{}) bool {
+	return src != nil && reflect.TypeOf(src).Kind() == reflect.Int
 }
