+++ cmd/kubeapps-apis/plugins/pkg/pkgutils/pkgutils_test.go
@@ -4,6 +4,9 @@
 package pkgutils
 
 import (
+	"bytes"
+	"encoding/json"
+	"reflect"
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
@@ -14,6 +17,7 @@ import (
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 	"helm.sh/helm/v3/pkg/chart"
+	structuralschema "k8s.io/apiextensions-apiserver/pkg/apiserver/schema"
 )
 
 const (
@@ -678,6 +682,395 @@ func TestSplitChartIdentifier(t *testing.T) {
 				}
 				if got, want := chartName, tc.chartName; !cmp.Equal(got, want) {
 					t.Errorf("mismatch (-want +got):\n%s", cmp.Diff(want, got))
+				}
+			}
+		})
+	}
+}
+
+func TestDefaultValuesFromSchema(t *testing.T) {
+	tests := []struct {
+		name           string
+		isCommentedOut bool
+		schema         []byte
+		expected       string
+		expectedErr    error
+	}{
+		{"schema with defaults", false, []byte(`properties:
+  valueWithDefault:
+    default: 80
+    description: Value with default
+    type: integer`,
+		),
+			`valueWithDefault: 80
+`, nil},
+		{"schema with without defaults (integer)", false, []byte(`properties:
+  missingDefaultInteger:
+    description: Missing default
+    type: integer`,
+		),
+			`missingDefaultInteger: 0
+`, nil},
+		{"schema with without defaults (number)", false, []byte(`properties:
+  missingDefaultNumber:
+    description: Missing default
+    type: number`,
+		),
+			`missingDefaultNumber: 0
+`, nil},
+		{"schema with without defaults (string)", false, []byte(`properties:
+  missingDefaultString:
+    description: Missing default
+    type: string`,
+		),
+			`missingDefaultString: ""
+`, nil},
+		{"schema with without defaults (boolean)", false, []byte(`properties:
+  missingDefaultBoolean:
+    description: Missing default
+    type: boolean`,
+		),
+			`missingDefaultBoolean: false
+`, nil},
+		{"schema with without defaults (array)", false, []byte(`properties:
+  missingDefaultArray:
+    description: Missing default
+    type: array`,
+		),
+			`missingDefaultArray: []
+`, nil},
+		{"schema with without defaults (object)", false, []byte(`properties:
+  missingDefaultObject:
+    description: Missing default
+    type: object`,
+		),
+			`missingDefaultObject: {}
+`, nil},
+		{"schema (mixed) with isCommentedOut=false", false, []byte(`properties:
+  missingDefaultObject:
+    description: Missing default
+    type: object
+  valueWithDefault:
+    default: 80
+    description: Value with default
+    type: integer`,
+		),
+			`missingDefaultObject: {}
+valueWithDefault: 80
+`, nil},
+		{"schema (mixed) with isCommentedOut=true", true, []byte(`properties:
+  missingDefaultObject:
+    description: Missing default
+    type: object
+  valueWithDefault:
+    default: 80
+    description: Value with default
+    type: integer`,
+		),
+			`# missingDefaultObject: {}
+# valueWithDefault: 80
+`, nil},
+		{"good schema (w/ additionalProperties: true, as per jsonschema draft 4)", true, []byte(`properties:
+  myAdditionalPropertiesProp:
+    type: object
+    additionalProperties: true
+`,
+		),
+			`# myAdditionalPropertiesProp: {}
+`, nil},
+		{"good schema (w/ additionalProperties: <schema>)", true, []byte(`properties:
+  myAdditionalPropertiesProp:
+    type: object
+    additionalProperties:
+      type: string
+`,
+		),
+			`# myAdditionalPropertiesProp: {}
+`, nil},
+		{"bad schema (w/ additionalProperties: string)", true, []byte(`properties:
+  myAdditionalPropertiesProp:
+    type: object
+    additionalProperties: string
+`,
+		),
+			`# myAdditionalPropertiesProp: {}
+`, nil},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			values, err := DefaultValuesFromSchema(tt.schema, tt.isCommentedOut)
+			if err != nil && tt.expectedErr == nil {
+				t.Errorf("unexpected error = %v", err)
+			}
+			if tt.expectedErr != nil {
+				if want, got := tt.expectedErr.Error(), err.Error(); !cmp.Equal(want, got) {
+					t.Errorf("in %s: mismatch (-want +got):\n%s", tt.name, cmp.Diff(want, got))
+				}
+			} else {
+				if want, got := tt.expected, values; !cmp.Equal(want, got) {
+					t.Errorf("mismatch in '%s': %s", tt.name, cmp.Diff(tt.expected, values))
+				}
+			}
+		})
+	}
+}
+
+// From https://github.com/kubernetes/apiextensions-apiserver/blob/release-1.21/pkg/apiserver/schema/defaulting/algorithm_test.go
+// With a new test case ("object without default values")
+func TestDefaultValues(t *testing.T) {
+	tests := []struct {
+		name     string
+		json     string
+		schema   *structuralschema.Structural
+		expected string
+	}{
+		{"empty", "null", nil, "null"},
+		{"scalar", "4", &structuralschema.Structural{
+			Generic: structuralschema.Generic{
+				Default: structuralschema.JSON{"foo"},
+			},
+		}, "4"},
+		{"scalar array", "[1,2]", &structuralschema.Structural{
+			Items: &structuralschema.Structural{
+				Generic: structuralschema.Generic{
+					Default: structuralschema.JSON{"foo"},
+				},
+			},
+		}, "[1,2]"},
+		{"object array", `[{"a":1},{"b":1},{"c":1}]`, &structuralschema.Structural{
+			Items: &structuralschema.Structural{
+				Properties: map[string]structuralschema.Structural{
+					"a": {
+						Generic: structuralschema.Generic{
+							Default: structuralschema.JSON{"A"},
+						},
+					},
+					"b": {
+						Generic: structuralschema.Generic{
+							Default: structuralschema.JSON{"B"},
+						},
+					},
+					"c": {
+						Generic: structuralschema.Generic{
+							Default: structuralschema.JSON{"C"},
+						},
+					},
+				},
+			},
+		}, `[{"a":1,"b":"B","c":"C"},{"a":"A","b":1,"c":"C"},{"a":"A","b":"B","c":1}]`},
+		// New test case checking our tweaks
+		{"object without default values", `{}`, &structuralschema.Structural{
+			Properties: map[string]structuralschema.Structural{
+				"a": {
+					Generic: structuralschema.Generic{
+						Type: "string",
+					},
+				},
+				"b": {
+					Generic: structuralschema.Generic{
+						Type: "boolean",
+					},
+				},
+				"c": {
+					Generic: structuralschema.Generic{
+						Type: "array",
+					},
+				},
+				"d": {
+					Generic: structuralschema.Generic{
+						Type: "number",
+					},
+				},
+				"e": {
+					Generic: structuralschema.Generic{
+						Type: "integer",
+					},
+				},
+				"f": {
+					Generic: structuralschema.Generic{
+						Type: "object",
+					},
+				},
+			},
+		}, `{
+  "a": "",
+  "b": false,
+  "c": [],
+  "d": 0,
+  "e": 0,
+  "f": {}
+}
+`},
+		{"object array object", `{"array":[{"a":1},{"b":2}],"object":{"a":1},"additionalProperties":{"x":{"a":1},"y":{"b":2}}}`, &structuralschema.Structural{
+			Properties: map[string]structuralschema.Structural{
+				"array": {
+					Items: &structuralschema.Structural{
+						Properties: map[string]structuralschema.Structural{
+							"a": {
+								Generic: structuralschema.Generic{
+									Default: structuralschema.JSON{"A"},
+								},
+							},
+							"b": {
+								Generic: structuralschema.Generic{
+									Default: structuralschema.JSON{"B"},
+								},
+							},
+						},
+					},
+				},
+				"object": {
+					Properties: map[string]structuralschema.Structural{
+						"a": {
+							Generic: structuralschema.Generic{
+								Default: structuralschema.JSON{"N"},
+							},
+						},
+						"b": {
+							Generic: structuralschema.Generic{
+								Default: structuralschema.JSON{"O"},
+							},
+						},
+					},
+				},
+				"additionalProperties": {
+					Generic: structuralschema.Generic{
+						AdditionalProperties: &structuralschema.StructuralOrBool{
+							Structural: &structuralschema.Structural{
+								Properties: map[string]structuralschema.Structural{
+									"a": {
+										Generic: structuralschema.Generic{
+											Default: structuralschema.JSON{"alpha"},
+										},
+									},
+									"b": {
+										Generic: structuralschema.Generic{
+											Default: structuralschema.JSON{"beta"},
+										},
+									},
+								},
+							},
+						},
+					},
+				},
+				"foo": {
+					Generic: structuralschema.Generic{
+						Default: structuralschema.JSON{"bar"},
+					},
+				},
+			},
+		}, `{"array":[{"a":1,"b":"B"},{"a":"A","b":2}],"object":{"a":1,"b":"O"},"additionalProperties":{"x":{"a":1,"b":"beta"},"y":{"a":"alpha","b":2}},"foo":"bar"}`},
+		{"empty and null", `[{},{"a":1},{"a":0},{"a":0.0},{"a":""},{"a":null},{"a":[]},{"a":{}}]`, &structuralschema.Structural{
+			Items: &structuralschema.Structural{
+				Properties: map[string]structuralschema.Structural{
+					"a": {
+						Generic: structuralschema.Generic{
+							Default: structuralschema.JSON{"A"},
+						},
+					},
+				},
+			},
+		}, `[{"a":"A"},{"a":1},{"a":0},{"a":0.0},{"a":""},{"a":"A"},{"a":[]},{"a":{}}]`},
+		{"null in nullable list", `[null]`, &structuralschema.Structural{
+			Generic: structuralschema.Generic{
+				Nullable: true,
+			},
+			Items: &structuralschema.Structural{
+				Properties: map[string]structuralschema.Structural{
+					"a": {
+						Generic: structuralschema.Generic{
+							Default: structuralschema.JSON{"A"},
+						},
+					},
+				},
+			},
+		}, `[null]`},
+		{"null in non-nullable list", `[null]`, &structuralschema.Structural{
+			Generic: structuralschema.Generic{
+				Nullable: false,
+			},
+			Items: &structuralschema.Structural{
+				Generic: structuralschema.Generic{
+					Default: structuralschema.JSON{"A"},
+				},
+			},
+		}, `["A"]`},
+		{"null in nullable object", `{"a": null}`, &structuralschema.Structural{
+			Generic: structuralschema.Generic{},
+			Properties: map[string]structuralschema.Structural{
+				"a": {
+					Generic: structuralschema.Generic{
+						Nullable: true,
+						Default:  structuralschema.JSON{"A"},
+					},
+				},
+			},
+		}, `{"a": null}`},
+		{"null in non-nullable object", `{"a": null}`, &structuralschema.Structural{
+			Properties: map[string]structuralschema.Structural{
+				"a": {
+					Generic: structuralschema.Generic{
+						Nullable: false,
+						Default:  structuralschema.JSON{"A"},
+					},
+				},
+			},
+		}, `{"a": "A"}`},
+		{"null in nullable object with additionalProperties", `{"a": null}`, &structuralschema.Structural{
+			Generic: structuralschema.Generic{
+				AdditionalProperties: &structuralschema.StructuralOrBool{
+					Structural: &structuralschema.Structural{
+						Generic: structuralschema.Generic{
+							Nullable: true,
+							Default:  structuralschema.JSON{"A"},
+						},
+					},
+				},
+			},
+		}, `{"a": null}`},
+		{"null in non-nullable object with additionalProperties", `{"a": null}`, &structuralschema.Structural{
+			Generic: structuralschema.Generic{
+				AdditionalProperties: &structuralschema.StructuralOrBool{
+					Structural: &structuralschema.Structural{
+						Generic: structuralschema.Generic{
+							Nullable: false,
+							Default:  structuralschema.JSON{"A"},
+						},
+					},
+				},
+			},
+		}, `{"a": "A"}`},
+		{"null unknown field", `{"a": null}`, &structuralschema.Structural{
+			Generic: structuralschema.Generic{
+				AdditionalProperties: &structuralschema.StructuralOrBool{
+					Bool: true,
+				},
+			},
+		}, `{"a": null}`},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			var in interface{}
+			if err := json.Unmarshal([]byte(tt.json), &in); err != nil {
+				t.Fatal(err)
+			}
+
+			var expected interface{}
+			if err := json.Unmarshal([]byte(tt.expected), &expected); err != nil {
+				t.Fatal(err)
+			}
+
+			defaultValues(in, tt.schema)
+			if !reflect.DeepEqual(in, expected) {
+				var buf bytes.Buffer
+				enc := json.NewEncoder(&buf)
+				enc.SetIndent("", "  ")
+				err := enc.Encode(in)
+				if err != nil {
+					t.Fatalf("unexpected result mashalling error: %v", err)
+				}
+				if tt.expected != buf.String() {
+					t.Errorf("mismatch (-want +got):\n%s", cmp.Diff(tt.expected, buf.String()))
 				}
 			}
 		})
